requirements.txt
flask-restful==0.3.5
flask==0.12
gunicorn==19.7.1
influxdb==4.1.0
liac-arff==2.1.0
pandas==0.20.1
pylru==1.0.9
requests==2.13.0
scikit-learn==0.18.1
https://github.com/nok/sklearn-porter/archive/v0.5.0.zip
